# JDBC: Optimize writeEntity calls