--- a/extension/persistence/relational-jdbc/src/main/java/org/apache/polaris/extension/persistence/relational/jdbc/JdbcBasePersistenceImpl.java
+++ b/extension/persistence/relational-jdbc/src/main/java/org/apache/polaris/extension/persistence/relational/jdbc/JdbcBasePersistenceImpl.java
@@ -93,7 +93,7 @@ public class JdbcBasePersistenceImpl implements BasePersistence, IntegrationPers
       boolean nameOrParentChanged,
       PolarisBaseEntity originalEntity) {
     try {
-      persistEntity(callCtx, entity, originalEntity, datasourceOperations);
+      persistEntity(callCtx, entity, originalEntity, datasourceOperations::executeUpdate);
     } catch (SQLException e) {
       throw new RuntimeException("Error persisting entity", e);
     }
@@ -111,7 +111,6 @@ public class JdbcBasePersistenceImpl implements BasePersistence, IntegrationPers
               PolarisBaseEntity entity = entities.get(i);
               PolarisBaseEntity originalEntity =
                   originalEntities != null ? originalEntities.get(i) : null;
-
               // first, check if the entity has already been created, in which case we will simply
               // return it.
               PolarisBaseEntity entityFound =
@@ -123,7 +122,7 @@ public class JdbcBasePersistenceImpl implements BasePersistence, IntegrationPers
                 // already been updated after the creation.
                 continue;
               }
-              persistEntity(callCtx, entity, originalEntity, statement);
+              persistEntity(callCtx, entity, originalEntity, statement::executeUpdate);
             }
             return true;
           });
@@ -139,12 +138,12 @@ public class JdbcBasePersistenceImpl implements BasePersistence, IntegrationPers
       @Nonnull PolarisCallContext callCtx,
       @Nonnull PolarisBaseEntity entity,
       PolarisBaseEntity originalEntity,
-      Object executor)
+      QueryAction queryAction)
       throws SQLException {
     ModelEntity modelEntity = ModelEntity.fromEntity(entity);
     if (originalEntity == null) {
       try {
-        execute(executor, generateInsertQuery(modelEntity, realmId));
+        queryAction.apply(generateInsertQuery(modelEntity, realmId));
       } catch (SQLException e) {
         if (datasourceOperations.isConstraintViolation(e)) {
           PolarisBaseEntity existingEntity =
@@ -172,7 +171,7 @@ public class JdbcBasePersistenceImpl implements BasePersistence, IntegrationPers
               "realm_id",
               realmId);
       try {
-        int rowsUpdated = execute(executor, generateUpdateQuery(modelEntity, params));
+        int rowsUpdated = queryAction.apply(generateUpdateQuery(modelEntity, params));
         if (rowsUpdated == 0) {
           throw new RetryOnConcurrencyException(
               "Entity '%s' id '%s' concurrently modified; expected version %s",
@@ -182,17 +181,6 @@ public class JdbcBasePersistenceImpl implements BasePersistence, IntegrationPers
         throw new RuntimeException(
             String.format("Failed to write entity due to %s", e.getMessage()), e);
       }
-    }
-  }
-
-  private int execute(Object executor, String query) throws SQLException {
-    if (executor instanceof Statement) {
-      // used for running in transaction
-      return ((Statement) executor).executeUpdate(query);
-    } else if (executor instanceof DatasourceOperations) {
-      return ((DatasourceOperations) executor).executeUpdate(query);
-    } else {
-      throw new IllegalArgumentException("Unsupported executor: " + executor);
     }
   }
 
@@ -929,5 +917,10 @@ public class JdbcBasePersistenceImpl implements BasePersistence, IntegrationPers
     PolarisStorageConfigurationInfo storageConfig =
         BaseMetaStoreManager.extractStorageConfiguration(callContext, entity);
     return storageIntegrationProvider.getStorageIntegrationForConfig(storageConfig);
+  }
+
+  @FunctionalInterface
+  private interface QueryAction {
+    Integer apply(String query) throws SQLException;
   }
 }