## About the change 

Optimize writeEntity calls by extracting the calls out of transaction. please ref to benchmarks below for the motivation for the change. 

**_Note : Handling transaction failures and retries are not in the scope of this pr. This is purely for optimization purpose._**


-----

## Running Branchmark : 

started the jdbc implementation backed Service : https://github.com/apache/polaris/pull/1470/files#diff-44e2226f0bc5712cf1cd511efd94242a08c9760b947809c7ebe76f121f480c76


Updated creds via 
```
--- a/benchmarks/src/gatling/resources/benchmark-defaults.conf
+++ b/benchmarks/src/gatling/resources/benchmark-defaults.conf
@@ -28,11 +28,11 @@ http {
 auth {
   # OAuth2 client ID for authentication
   # Required: Must be provided in configuration
-  client-id = null
+  client-id = root

   # OAuth2 client secret for authentication
   # Required: Must be provided in configuration
-  client-secret = null
+  client-secret = s3cr3t
 }
```

go to benchmark tools and run the following benchmarks : 

```
# Dataset creation
./gradlew gatlingRun --simulation org.apache.polaris.benchmarks.simulations.CreateTreeDataset

# Read/Update operations
./gradlew gatlingRun --simulation org.apache.polaris.benchmarks.simulations.ReadUpdateTreeDataset

# Read-only operations
./gradlew gatlingRun --simulation org.apache.polaris.benchmarks.simulations.ReadTreeDataset
```


#### withRunInTransaction

```
========================================================================================================================
---- Global Information -------------------------------------------------------------|---Total---|-----OK----|----KO----
> request count                                                                      |    30,108 |    30,092 |        16
> min response time (ms)                                                             |         1 |         1 |         7
> max response time (ms)                                                             |       603 |       603 |       243
> mean response time (ms)                                                            |        17 |        17 |        99
> response time std deviation (ms)                                                   |        23 |        23 |        74
> response time 25th percentile (ms)                                                 |         6 |         6 |        44
> response time 50th percentile (ms)                                                 |        11 |        11 |        76
> response time 75th percentile (ms)                                                 |        22 |        22 |       171
> response time 99th percentile (ms)                                                 |        96 |        94 |       243
> mean throughput (rps)                                                              |      83.4 |     83.36 |      0.04
---- Response Time Distribution ----------------------------------------------------------------------------------------
> OK: t < 800 ms                                                                                         30,092 (99.95%)
> OK: 800 ms <= t < 1200 ms                                                                                   0     (0%)
> OK: t >= 1200 ms                                                                                            0     (0%)
> KO                                                                                                         16  (0.05%)
---- Errors ------------------------------------------------------------------------------------------------------------
> status.find.is(200), but actually found 500                                                                15 (93.75%)
> status.find.is(204), but actually found 500                                                                 1  (6.25%)
========================================================================================================================
```

### without runInTransaction

```
========================================================================================================================
---- Global Information -------------------------------------------------------------|---Total---|-----OK----|----KO----
> request count                                                                      |    30,009 |    30,001 |         8
> min response time (ms)                                                             |         1 |         1 |        10
> max response time (ms)                                                             |       293 |       293 |        89
> mean response time (ms)                                                            |        16 |        16 |        33
> response time std deviation (ms)                                                   |        16 |        16 |        27
> response time 25th percentile (ms)                                                 |         6 |         6 |        16
> response time 50th percentile (ms)                                                 |        12 |        12 |        20
> response time 75th percentile (ms)                                                 |        23 |        23 |        60
> response time 99th percentile (ms)                                                 |        77 |        76 |        89
> mean throughput (rps)                                                              |     83.13 |     83.11 |      0.02
---- Response Time Distribution ----------------------------------------------------------------------------------------
> OK: t < 800 ms                                                                                         30,001 (99.97%)
> OK: 800 ms <= t < 1200 ms                                                                                   0     (0%)
> OK: t >= 1200 ms                                                                                            0     (0%)
> KO                                                                                                          8  (0.03%)
---- Errors ------------------------------------------------------------------------------------------------------------
> status.find.is(200), but actually found 500                                                                 6    (75%)
> status.find.is(204), but actually found 500                                                                 2    (25%)
========================================================================================================================
```