@singhpk234 2025-04-29
@dimas-b here is the logs for he same : 
```
2025-04-29 20:35:59.540 UTC [1] LOG:  listening on Unix socket "/var/run/postgresql/.s.PGSQL.5432"
2025-04-29 20:35:59.545 UTC [63] LOG:  database system was shut down at 2025-04-29 20:35:59 UTC
2025-04-29 20:35:59.559 UTC [1] LOG:  database system is ready to accept connections
2025-04-29 20:39:14.685 UTC [175] ERROR:  could not serialize access due to read/write dependencies among transactions
2025-04-29 20:39:14.685 UTC [175] DETAIL:  Reason code: Canceled on conflict out to pivot 894, during read.
2025-04-29 20:39:14.685 UTC [175] HINT:  The transaction might succeed if retried.
2025-04-29 20:39:14.685 UTC [175] STATEMENT:  SELECT entity_version, to_purge_timestamp, internal_properties, catalog_id, purge_timestamp, sub_type_code, create_timestamp, last_update_timestamp, parent_id, name, id, drop_timestamp, properties, grant_records_version, type_code FROM POLARIS_SCHEMA.ENTITIES WHERE (catalog_id, id) IN ((0, 8213068413194098796),(0, 2285716631442680836),(0, 3182916641487612758),(3182916641487612758, 3957433893749870399),(0, 0),(3182916641487612758, 6835667964051947808)) AND realm_id = 'POLARIS'
2025-04-29 20:39:14.688 UTC [84] ERROR:  could not serialize access due to read/write dependencies among transactions
2025-04-29 20:39:14.688 UTC [84] DETAIL:  Reason code: Canceled on identification as a pivot, during commit attempt.
2025-04-29 20:39:14.688 UTC [84] HINT:  The transaction might succeed if retried.
2025-04-29 20:39:14.688 UTC [84] STATEMENT:  COMMIT
2025-04-29 20:39:14.869 UTC [173] ERROR:  could not serialize access due to read/write dependencies among transactions
```