@dimas-b 2025-04-30
@singhpk234 : could you re-run the perf. tests on the final state of this PR and re-compare with `main` before this change?